Add delete method to interface and add some checks before deletion

+++ private/entities/Recipe/Recipe.php
@@ -118,10 +118,19 @@ class Recipe implements RecipeInterface, DbObjectInterface, \JsonSerializable {
 
   public function delete() : bool {
     global $Controller;
-    $query = new QueryBuilder(EQueryType::qtDELETE, 'recipes');
-    $query->where('recipes', 'recipe_id', '=', intval($this->recipe_id))
-          ->limit(1);
-    return $Controller->delete($query);
+    if (
+      $Controller->isAuthenticated() &&
+      (
+        (is_null($this->user_id) && $Controller->User()->isAdmin()) ||
+        (!is_null($this->user_id) && ($this->user_id == $Controller->User()->getId() || $Controller->User()->isAdmin()))
+      )
+    ) {
+      $query = new QueryBuilder(EQueryType::qtDELETE, 'recipes');
+      $query->where('recipes', 'recipe_id', '=', intval($this->recipe_id))
+            ->limit(1);
+      return $Controller->delete($query);
+    }
+    return false;
   }
 
   public function getCookedCount() : int {
